services/idp/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-jose[cryptography]==3.3.0
cryptography==42.0.5
PyYAML==6.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
aiosqlite==0.20.0
redis[asyncio]==5.0.4
httpx==0.27.0
python-ulid==2.2.0
prometheus-client==0.20.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
